The ml-ane-transformers repo provides a reference implementation of a channels-first layer norm [here](https://github.com/apple/ml-ane-transformers/blob/da64000fa56cc85b0859bc17cb16a3d753b8304a/ane_transformers/reference/layer_norm.py). This adds a pass to replace that with an equivalent MIL `layer_norm` op.

This uses less ops (7 or 9 depending on if the elementwise affine is used) and makes the MIL/graph easier to read. I mostly followed the pattern of the other similar passes, happy to make any adjustments!