@aseemw 2023-09-14
@smpanaro thanks for the PR. 
Can you please also test the impact of this pass on the latency of the model? That is, taking the distill bert model from the transformers repo, and converting it with main branch, checking its latency on the phone with Xcode perf tab and repeating that experiment with conversion from this PR.