@smpanaro 2023-09-14
Sure thing! I have iOS 16.6.1. Would that work? I can also run tests on macOS Ventura and Sonoma if needed.