@smpanaro 2023-09-15
For the latency tests I compared with feed1741 since that's where I branched from (seems like the only changes since then are doc-related).

- Time is Xcode's median from the first run of each OS/SHA combo.
- Xcode 15 beta for Sonoma. Xcode 14.3.1 otherwise.

|OS|main feed1741|PR ff8f8994|
|--|--|--|
|iOS 16.6.1|10.23ms|10.26ms|
|macOS 13.5.1|8.86ms|8.82ms|
|macOS 14 23A5328b*|8.91ms|8.99ms|

<sub>* pretty sure, might be off by one beta</sub>

<details>
<summary>Xcode screenshots</summary>

|OS|main feed1741|PR ff8f8994|
|--|--|--|
|iOS 16.6.1|<img width="1092" alt="feed1741-iphone" src="https://github.com/apple/coremltools/assets/2950214/b6ed3472-1d94-410f-a1fe-4b8fbf1ffa55">|<img width="1095" alt="ff8f8994-iphone" src="https://github.com/apple/coremltools/assets/2950214/30e880f0-9aae-46df-801e-431d4d08426b">|
|macOS 13.5.1|<img width="1092" alt="feed1741-ventura" src="https://github.com/apple/coremltools/assets/2950214/38b6fc35-2b67-4c55-b7b8-62e2421f4519">|<img width="1095" alt="ff8f8994-ventura" src="https://github.com/apple/coremltools/assets/2950214/1b4252fd-0112-4998-ab4b-ff7cb3b8f819">|
|macOS 14 23A5328b*|<img width="1096" alt="feed1741-sonoma-median" src="https://github.com/apple/coremltools/assets/2950214/aee40196-11ac-48a7-b7d6-4eb117b6e347">|<img width="1095" alt="ff8f8994-sonoma-median" src="https://github.com/apple/coremltools/assets/2950214/40f988ee-c5de-4f5d-b77d-02a4820ef8ae">|

<sub>* pretty sure, might be off by one beta</sub>
</details>

<details>
<summary>Netron screenshots</summary >

|main feed1741|<img width="1186" alt="netron-feed1741" src="https://github.com/apple/coremltools/assets/2950214/a4c94749-5996-4fc7-9ecd-ea86a11a6ddd">|
|--|--|
|PR ff8f8994|<img width="1186" alt="netron-ff8f8994" src="https://github.com/apple/coremltools/assets/2950214/9a0d36f4-9d61-499d-abc2-1040cee56ec6">|

</details>


Can find somewhere to upload the mlpackages if it's helpful. I created them by saving at this line in the [distilbert test](https://github.com/apple/ml-ane-transformers/blob/da64000fa56cc85b0859bc17cb16a3d753b8304a/ane_transformers/huggingface/test_distilbert.py#L142). Also, since it was printed in the process the PSNR is comparable: 94.99243529711 for main, 94.32243613597612 for PR.

Will take a look at the PR comments next.