@TobyRoseman 2023-10-09
@smpanaro - thank you for this contribution. I've tested this change internally on a number of different devices. The results look good. If you can address the most recent feedback from @jakesabathia2 and we can get a clean CI run, I will merge this pull request.