@TobyRoseman 2023-10-23
Updated CI: https://gitlab.com/coremltools1/coremltools/-/pipelines/1046860640