@aseemw 2023-10-24
> https://gitlab.com/coremltools1/coremltools/-/pipelines/1046860640

Thanks the pipeline is green! Should we rebase and merge the change?